Refactor:   - Rename Askfm.DLL -> Askfm.DAL (Which is Data Acess Layer)   - Moving Enums Notification & Thread Status to the Write Place Data Access Layer
Fix:
  - Edit References for each project
    - API have reference of (BLL,Shared)
    - BLL have reference of (DAL,Shared)
    - DAL have reference of (Shared)
