+++ AskFm/AskFm.DAL/Enums/NotificationStatus.cs
@@ -1,0 +1,11 @@
+namespace AskFm.DAL.Enums;
+
+public enum NotificationStatus
+{
+    ANSWER = 1,
+    QUESTION = 2,
+    FOLLOW = 3,
+    COMMENT_LIKE = 4,
+    QUESTION_LIKE = 5,
+    REPLAY = 6
+}
