+++ AskFm/AskFm.DAL/Enums/ThreadStatus.cs
@@ -1,0 +1,6 @@
+namespace AskFm.DAL.Enums;
+public enum ThreadStatus
+{
+    PRIVATE = 1,
+    PUBLIC = 2
+}
