+++ AskFm/AskFm.DAL/Moodels/Comment.cs
@@ -1,6 +1,4 @@
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace AskFm.DLL.Moodels;
+namespace AskFm.DAL.Moodels;
 
 public class Comment
 {
@@ -8,16 +6,16 @@ public class Comment
     public string Content { get; set; }
     public DateTime CreatedAt { get; set; }
     public int LikeCount { get; set; }
-    
+
     public int UserId { get; set; }
     public User User { get; set; }
-    
+
     public int ThreadId { get; set; }
     public Thread Thread { get; set; }
-    
+
     public int? ParentCommentId { get; }
     public Comment ParentComment { get; set; }
-    
+
     public ICollection<Comment> Replies { get; set; }
-    public ICollection<CommentLike>  CommentLikes { get; set; }
+    public ICollection<CommentLike> CommentLikes { get; set; }
 }
