+++ AskFm/AskFm.DAL/Moodels/Notification.cs
@@ -1,12 +1,11 @@
-namespace AskFm.DLL.Moodels;
+namespace AskFm.DAL.Moodels;
 
 public class Notification
 {
+    public GCNotificationStatus Type;
     public int Id { get; set; }
     public int UserId { get; set; }
     public User User { get; set; }
-
-    public GCNotificationStatus Type;
     public bool isRead { get; set; }
     public int ResourceId { get; set; }
     public string jsonContent { get; set; }
