+++ AskFm/AskFm.DAL/Moodels/ThreadLike.cs
@@ -1,12 +1,12 @@
-namespace AskFm.DLL.Moodels;
+namespace AskFm.DAL.Moodels;
 
 public class ThreadLike
 {
     public int ThreadId { get; set; }
     public Thread Thread { get; set; }
-    
+
     public int UserId { get; set; }
     public User User { get; set; }
-    
+
     public DateTime CreatedAt { get; set; }
 }
