--- AskFm/AskFm.DLL/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace AskFm.DLL
-{
-    public class Class1
-    {
-
-    }
-}
