+++ AskFm/AskFm.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AskFm.API", "AskFm.API\AskF
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AskFm.BLL", "AskFm.BLL\AskFm.BLL.csproj", "{D659562C-FEF7-486F-A320-4FCE8EFD6202}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AskFm.DLL", "AskFm.DLL\AskFm.DLL.csproj", "{DABC5408-AC02-42E5-B4D7-9CE6FA449BA6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AskFm.DAL", "AskFm.DAL\AskFm.DAL.csproj", "{DABC5408-AC02-42E5-B4D7-9CE6FA449BA6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{CE580260-EB9E-4AFE-8ED5-34EF36BD5006}"
 EndProject
