--- AskFm/Shared/NotificationStatus.cs
@@ -1,11 +1,0 @@
-namespace Shared;
-
-public enum NotificationStatus
-{
-    ANSWER,
-    QUESTION,
-    FOLLOW,
-    COMMENT_LIKE,
-    QUESTION_LIKE,
-    REPLAY
-}
