--- AskFm/Shared/ThreadStatus.cs
@@ -1,7 +1,0 @@
-namespace Shared;
-
-public enum ThreadStatus
-{
-    PRIVATE,
-    PUBLIC,
-}
